Early return for bidirectional vergesort when the data is already sorted
One of the first functions called by the bidirectional version of
vergesor is is_sorted_until since it easily allows to find the length of
the first ascending run. This commit uses the information to check
whether the collection is already sorted with a single cheap iterator
comparison and makes vergesort return early when it is the case.