Remove overly generic calls to small helpers
This commit replaces the calls to std::distance and std::advance by
their underlying operations when called on types known to satisfy the
RandomAccessIterator named requirement. This ensures that the error
message points directly to the relevant algorithm when passed a
different iterator instead of in the implementation of a standard
library helper function.